factor out tsProject and its dependencies
