+++ package.json
@@ -52,6 +52,7 @@
     "ui/tournamentSchedule",
     "ui/tournamentCalendar",
     "ui/tree",
+    "ui/@build/tsProject",
     "ui/@types/lichess"
   ]
 }
