+++ ui/challenge/gulpfile.js
@@ -1,2 +1,2 @@
-require('../gulp/tsProject.js')('LichessChallenge', 'lichess.challenge', __dirname);
+require('@build/tsProject')('LichessChallenge', 'lichess.challenge', __dirname);
 require('../gulp/cssProject.js')(__dirname);
