+++ ui/dasher/package.json
@@ -8,18 +8,9 @@
     "snabbdom": "ornicar/snabbdom#0.7.1-lichess"
   },
   "devDependencies": {
+    "@build/tsProject": "^2",
     "@types/jquery": "^2.0",
     "@types/jqueryui": "^1.11",
-    "@types/lichess": "1.0.0",
-    "browserify": "^16",
-    "fancy-log": "^1",
-    "gulp": "^4",
-    "gulp-size": "^3",
-    "gulp-terser": "^1",
-    "tsify": "^4",
-    "typescript": "^3",
-    "vinyl-source-stream": "^2",
-    "vinyl-buffer": "^1",
-    "watchify": "^3"
+    "@types/lichess": "1.0.0"
   }
 }
