+++ ui/notify/gulpfile.js
@@ -1,3 +1,3 @@
-require('../gulp/tsProject.js')('LichessNotify', 'lichess.notify', __dirname);
+require('@build/tsProject')('LichessNotify', 'lichess.notify', __dirname);
 
 require('../gulp/cssProject.js')(__dirname);
