+++ ui/serviceWorker/gulpfile.js
@@ -1,3 +1,3 @@
-const lilaGulp = require('../gulp/tsProject.js');
+const lilaGulp = require('@build/tsProject');
 
 lilaGulp(undefined, 'lichess.serviceWorker', __dirname);
