+++ ui/serviceWorker/package.json
@@ -5,15 +5,6 @@
   "author": "Thibault Duplessis",
   "license": "AGPL-3.0",
   "devDependencies": {
-    "browserify": "^16",
-    "fancy-log": "^1",
-    "gulp": "^4",
-    "gulp-size": "^3",
-    "gulp-terser": "^1",
-    "tsify": "^4",
-    "typescript": "^3",
-    "vinyl-source-stream": "^2",
-    "vinyl-buffer": "^1",
-    "watchify": "^3"
+    "@build/tsProject": "^2"
   }
 }
