+++ ui/tournament/gulpfile.js
@@ -1,3 +1,3 @@
-require('../gulp/tsProject.js')('LichessTournament', 'lichess.tournament', __dirname);
+require('@build/tsProject')('LichessTournament', 'lichess.tournament', __dirname);
 
 require('../gulp/cssProject.js')(__dirname);
