+++ ui/tournamentCalendar/gulpfile.js
@@ -1,3 +1,3 @@
-const lilaGulp = require('../gulp/tsProject.js');
+const lilaGulp = require('@build/tsProject');
 
 lilaGulp('LichessTournamentCalendar', 'lichess.tournamentCalendar', __dirname);
